pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = "Work TODO and project management"
authors = ["Constantino Schillebeeckx <constantino.schillebeeckx@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
python-decouple = "^3.6"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
flake8 = "^4.0.1"
black = "^22.6.0"
autopep8 = "^1.6.0"
isort = "^5.10.1"
types-requests = "^2.28.1"
pre-commit = "^2.20.0"
toml = "^0.10.2"
autoflake = "^1.4"
autoimport = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 115

[tool.isort]
profile = "black"
atomic = true
line_length = 115
indent = '    '
force_alphabetical_sort_within_sections = true
use_parentheses = true

[tool.autopep8]
max_line_length = 115

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
